@@ -2,23 +2,20 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
 use Illuminate\View\View;
 use Illuminate\Http\Request;
 
-class HomeController extends Controller
+class HomeController extends Controller implements HasMiddleware
 {
 
 
-    /**
-     * Create a new controller instance.
-     *
-     * @return void
-     */
-    public function __construct()
+    public static function middleware(): array
     {
-        $this->middleware('auth');
+        return ['auth'];
 
-    }//end __construct()
+    }//end middleware()
 
 
     /**
