@@ -8,7 +8,8 @@ interface LibrisInterface
 
 
     // public function __construct()
-    public function addDocument($file, $log = false);
+    public function addDocument($file, $log=false);
+
 
     // public function indexDocument(ParserService $parser)
     // public function getParser($file)
@@ -25,11 +26,19 @@ interface LibrisInterface
     // function countAllPages($docId = false){
     // function fetchAllPages($docId = false, $size = 100, $searchAfter = false){
     public function systems();
+
+
     public function docsBySystem($system, $page, $perpage, $tag);
+
+
     // public static function tagSort($a, $b)
     public function tagsForSystem($system);
+
+
     // public function searchPages($terms, $system, $document, $tag, $page = 1, $size = 60)
-    public function getDocThumbnail($doc, $regen = false);
+    public function getDocThumbnail($doc, $regen=false);
+
+
     // public function updateDocThumbnail($doc)
     // protected function saveDocThumbnail(\Imagick $image, $file)
     // public function generateDocThumbnail($file)
