@@ -43,11 +43,12 @@ class User extends Authenticatable
 
     }//end casts()
 
+
     protected function getIsAdminAttribute(): bool
     {
         return true;
 
-    }//end checkAdmin()
+    }//end getIsAdminAttribute()
 
 
 }//end class
